loader: Fix skipping all layers if one bad is found
The logic for loaderScanForLayers would jump to out if any issue was found
loading any layer. This prevented any correct layers from being detected.
The other fix is in loader_get_json to return if the length is zero,
which would cause a VK_ERROR_OUT_OF_HOST_MEMORY to be thrown because
cJSON_Parse would fail.

Change-Id: Ie7a5de5d0f38edd9b9012b9cd22fecd9ed86eb5c