build: Fix `loader_genvk.py` error logging
A modification to one line of code was missed when updating the loader
genvk file for the 1.2.140 header update and when the change was
incorporated it exposed the newly added `logDiag` function which can be
found in `reflib.py` in Vulkan-Docs. Instead of adding another script to
the Loader repo, `write` functionality was reinstated in
`loader_genvk.py`.

Change-Id: Idf792b24c86120a8c278f46ac8a47013c4c37f16