loader: Shift around vk_loader_platform headers
Moved the includes into their own block, instead of living in the block for
each operating system. This way is easier to see what includes are in the file
in its entirety.

Also, _GNU_SOURCE was uncommented since it appears in each source file
individually and is needed for alloca and secure_getenv